Adds check error to resource types and resource configs
In radar, sets the check error on the resource config for cases after it
is created. A check error is also now set on the resource type from
radar. The api now displays the check error on the
resource/resource_type as the check setup error and the check error on
the resource config as the check error.

concourse/concourse#2386