switches scheduler and exec to use new resource config versions table
Implements the functionality needed for having the scheduler and exec
packages to use the new resource config versions (rather than the old
versioned resources). Most of this work is also for result of changing
the BuildInput object to not include a VersionedResource object and
instead only hold the version and resource config version id. These
changes involve a change in the API for public builds, which will need
some work to be done in web to accept the new outputs.

concourse/concourse#2386

Signed-off-by: Mark Huang <mhuang@pivotal.io>